[SSN-77][JQ/ES] Use PublicProjectListing on homepage than a new component.
